extracts layouts into separate module
This commit really clarifies the layout system. A layout is an
association list of UIs, where the key is used to tag messages to make
sure they go to the correct child. All the logic associated with
composing UI elements is isolated in this module, dramatically
reducing boilerplate code. To create a new layout, you only need to
provide a function that takes the association list of rendered
children, and render them together. Register this function in the
layout module, and you're all set.