big NewRideButton reorg
The NewRideButton can now be configured to use different default
values per field. This required creating a way to specify a template
that default data could be plugged into. I did this by creating a
lexical environment that is passed down into the parsers, so that the
target panes can piece them together. This environment could be useful
for a lot of other things, but that's a subject for another commit.